1. Renames Rollsite to RollSite. 2. Renames roll_site.py under pair_store to avoid ambiguity.
