+++ python/eggroll/core/pair_store/__init__.py
@@ -32,8 +32,8 @@ def create_pair_adapter(options: dict):
     elif options["store_type"] == StoreTypes.ROLLPAIR_CACHE:
         ret = CacheAdapter(options=options)
     elif options["store_type"] == StoreTypes.ROLLPAIR_ROLLSITE:
-        from eggroll.core.pair_store.roll_site import RollsiteAdapter
-        ret = RollsiteAdapter(options=options)
+        from eggroll.core.pair_store.roll_site_adapter import RollSiteAdapter
+        ret = RollSiteAdapter(options=options)
     else:
         raise NotImplementedError(options)
     return ret
